langgraph/requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0

# Additional dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0

# Optional: For local LLM support
# ollama>=0.1.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0 